move SetStackEffect to a separate file
