+++ lib/registerTypes/RegisterTypes.h
@@ -13,6 +13,7 @@
 #include "../networkPacks/PacksForClientBattle.h"
 #include "../networkPacks/PacksForServer.h"
 #include "../networkPacks/PacksForLobby.h"
+#include "../networkPacks/SetStackEffect.h"
 #include "../VCMI_Lib.h"
 #include "../CArtHandler.h"
 #include "../CCreatureSet.h"
