+++ lib/spells/BattleSpellMechanics.cpp
@@ -17,6 +17,7 @@
 #include "../battle/IBattleState.h"
 #include "../battle/CBattleInfoCallback.h"
 #include "../networkPacks/PacksForClientBattle.h"
+#include "../networkPacks/SetStackEffect.h"
 #include "../CStack.h"
 
 VCMI_LIB_NAMESPACE_BEGIN
