+++ lib/spells/effects/Clone.cpp
@@ -16,6 +16,7 @@
 #include "../../battle/IBattleState.h"
 #include "../../battle/CUnitState.h"
 #include "../../networkPacks/PacksForClientBattle.h"
+#include "../../networkPacks/SetStackEffect.h"
 #include "../../serializer/JsonSerializeFormat.h"
 
 VCMI_LIB_NAMESPACE_BEGIN
