+++ lib/spells/effects/Dispel.cpp
@@ -22,6 +22,7 @@
 #include "../../battle/Unit.h"
 #include "../../bonuses/BonusList.h"
 #include "../../networkPacks/PacksForClientBattle.h"
+#include "../../networkPacks/SetStackEffect.h"
 #include "../../serializer/JsonSerializeFormat.h"
 
 VCMI_LIB_NAMESPACE_BEGIN
