+++ test/spells/effects/EffectFixture.cpp
@@ -14,6 +14,7 @@
 #include <vstd/RNG.h>
 
 #include "../../../lib/networkPacks/PacksForClientBattle.h"
+#include "../../../lib/networkPacks/SetStackEffect.h"
 
 #include "../../../lib/serializer/JsonDeserializer.h"
